fix: all of the external methods should have callThroughEVC modifier
from evc docs: to avoid status-check-related issues, it is advised to use the call function in the callback manner (meaning a Vault can call back itself using call) when the Vault is called directly. It will ensure the Account and Vault Status Checks are always deferred, at least until the end of the call.