audit-fix: in normalizedToFull, full should refer to current balance of user, not the totalSupply
- The full USD balance: balanceOf(violator) (the total USD worth of each enabled `tokenId`)
- The requested USD amount: transfer amount

So, for each tokenId, (the current ERC6909 balance of the violator* requested USD amount) / full USD balance should be transferred to the liquidator. This ensures that the liquidator receives a proportional share of each tokenId, and the total value equals the requested USD amount.

Earlier, total supply of tokenId (regardless of what the current balance is) * requested USD amount / full USD balance was being used, which is incorrect if violator doesn't have the total supply of tokenId