feat: support v4 pools with yield harvesting hook
the only differnce is that now we are not calculating the pending fees because it wouldn't have been accurate due to pending yield. We let the contract have the actual fees after decreasing liquidity. We know how much principal user should have gotten so the rest of it must be the fees. Except this time it's the most updated fees because the hook would have harvested yiled before removing liquidity